revert did context
