+++ src/App.tsx
@@ -12,7 +12,6 @@ import 'react-toastify/dist/ReactToastify.css';
 import { useLocalStorage } from 'usehooks-ts';
 import { BiometricProvider } from './contexts/BiometricContext';
 import { DarkModeContext } from './contexts/DarkModeContext';
-import { DidProvider } from './contexts/DidContext';
 import { ErrorProvider } from './contexts/ErrorContext';
 import {
   getBrowserLanguage,
@@ -176,15 +175,13 @@ function AppInner() {
     isLocaleInitialized && (
       <I18nProvider i18n={i18n}>
         <WalletProvider>
-          <DidProvider>
-            <PeerProvider>
-              <WalletConnectProvider>
-                <PriceProvider>
-                  <RouterProvider router={router} />
-                </PriceProvider>
-              </WalletConnectProvider>
-            </PeerProvider>
-          </DidProvider>
+          <PeerProvider>
+            <WalletConnectProvider>
+              <PriceProvider>
+                <RouterProvider router={router} />
+              </PriceProvider>
+            </WalletConnectProvider>
+          </PeerProvider>
         </WalletProvider>
       </I18nProvider>
     )
